big code move around; remove AP_ prefix from source files and consolidate header files (I may well break these up a bit more in a future commit).
--HG--
rename : src/AP_AtomDefinitions.h => src/AtomDefs.h
rename : src/AP_CDTOC.cpp => src/CDtoc.cpp
rename : src/AP_CDTOC.h => src/CDtoc.h
rename : src/AP_arrays.cpp => src/arrays.cpp
rename : src/APar_zlib.cpp => src/compress.cpp
rename : src/AP_AtomExtracts.cpp => src/extracts.cpp
rename : src/AP_iconv.cpp => src/iconv.cpp
rename : src/AP_ID3v2_tags.cpp => src/id3v2.cpp
rename : src/AP_ID3v2_tags.h => src/id3v2.h
rename : src/AP_ID3v2_FrameDefinitions.h => src/id3v2defs.h
rename : src/AP_ID3v2_Definitions.h => src/id3v2types.h
rename : src/AP_MetadataListings.cpp => src/metalist.cpp
rename : src/AP_NSFile_utils.mm => src/nsfile.mm
rename : src/AP_NSImage.mm => src/nsimage.mm
rename : src/AtomicParsley.cpp => src/parsley.cpp
rename : src/APar_sha1.cpp => src/sha1.cpp
rename : src/AP_commons.cpp => src/util.cpp
rename : src/AP_commons.h => src/util.h
rename : src/APar_uuid.cpp => src/uuid.cpp
