+++ src/AtomicParsley.h
@@ -105,218 +105,18 @@
 #define strncasecmp strnicmp
 #endif
 
-#include "AP_commons.h"
-#include "AP_ID3v2_Definitions.h"
+#include "util.h"
 
-const uint32_t AtomFlags_Data_Binary = 0;           // Atom version 1byte/ Atom flags 3 bytes; 0x00 00 00 00
-const uint32_t AtomFlags_Data_Text = 1;             // UTF-8, no termination
-const uint32_t AtomFlags_Data_JPEGBinary = 13;      // \x0D
-const uint32_t AtomFlags_Data_PNGBinary = 14;       // \x0E
-const uint32_t AtomFlags_Data_UInt = 21;            // \x15 for cpil, tmpo, rtng; iTMS atoms: cnID, atID, plID, geID, sfID, akID
-const uint32_t AtomFlags_Data_uuid_binary = 88;     // 0x58 for uuid atoms that contain files
+#define MAX_ATOMS 1024
+#define MAXDATA_PAYLOAD 1256
 
-enum {
-  UTF8_iTunesStyle_256glyphLimited        = 0, //no NULL termination
-  UTF8_iTunesStyle_Unlimited              = 1, //no NULL termination
-	UTF8_iTunesStyle_Binary                 = 3, //no NULL termination, used in purl & egid
-  UTF8_3GP_Style                          = 8, //terminated with a NULL uint8_t
-  UTF16_3GP_Style                         = 16 //terminated with a NULL uint16_t
-};
+#define DEFAULT_PADDING_LENGTH          2048;
+#define MINIMUM_REQUIRED_PADDING_LENGTH 0;
+#define MAXIMUM_REQUIRED_PADDING_LENGTH 5000;
 
-enum {
-	UNDEFINED_STYLE                         = 0,
-	ITUNES_STYLE                            = 100,
-	THIRD_GEN_PARTNER                       = 300, //3gpp files prior to 3gp6
-	THIRD_GEN_PARTNER_VER1_REL6             = 306, //3GPP Release6 the first spec to contain the complement of assets
-	THIRD_GEN_PARTNER_VER1_REL7             = 307, //3GPP Release7 introduces ID32 atoms
-	THIRD_GEN_PARTNER_VER2                  = 320, //3gp2 files
-	THIRD_GEN_PARTNER_VER2_REL_A            = 321, //3gp2 files, 3GPP2 C.S0050-A introduces 'gadi'
-	MOTIONJPEG2000                          = 400
-};
+#include "ap_types.h"
 
-struct AtomicInfo  {
-	short     AtomicNumber;
-	uint64_t  AtomicStart;
-	uint64_t  AtomicLength;
-	uint64_t  AtomicLengthExtended;
-	char*     AtomicName;
-	char*			ReverseDNSname;
-	char*			ReverseDNSdomain;
-	uint8_t   AtomicContainerState;
-	uint8_t   AtomicClassification; 
-	uint32_t	AtomicVerFlags;					//used by versioned atoms and derivatives
-	uint16_t  AtomicLanguage;					//used by 3gp assets & ID32 atoms only
-	uint8_t   AtomicLevel;
-	char*     AtomicData;
-	int       NextAtomNumber;					//our first atom is numbered 0; the last points back to it - so watch it!
-	uint32_t  ancillary_data;         //just contains a simple number for atoms that contains some interesting info (like stsd codec used)
-	uint8_t   uuid_style;
-	char*     uuid_ap_atomname;
-	ID3v2Tag* ID32_TagInfo;
-};
-
-//currently this is only used on Mac OS X to set type/creator for generic '.mp4' file extension files. The Finder 4 character code TYPE is what determines whether a file appears as a video or an audio file in a broad sense.
-struct EmployedCodecs {
-	bool has_avc1;
-	bool has_mp4v;
-	bool has_drmi;
-	bool has_alac;
-	bool has_mp4a;
-	bool has_drms;
-	bool has_timed_text; //carries the URL - in the mdat stream at a specific time - thus it too is timed.
-	bool has_timed_jpeg; //no idea of podcasts support 'png ' or 'tiff'
-	bool has_timed_tx3g; //this IS true timed text stream
-	bool has_mp4s; //MPEG-4 Systems
-	bool has_rtp_hint; //'rtp '; implies hinting
-};
-
-enum {
-	MEDIADATA__PRECEDES__MOOV = 2,
-  ROOT_META__PRECEDES__MOOV = 4,
-	MOOV_META__PRECEDES__TRACKS = 8,
-	MOOV_UDTA__PRECEDES__TRACKS = 16,
-	
-	PADDING_AT_EOF = 0x1000000
-};
-
-struct FreeAtomListing {
-	AtomicInfo* free_atom;
-	FreeAtomListing* next_free_listing;
-};
-
-struct DynamicUpdateStat  {
-	bool updage_by_padding;
-	bool reorder_moov;
-	bool moov_was_mooved;
-	bool prevent_dynamic_update;
-
-	uint32_t optimization_flags;
-	
-	uint64_t padding_bytes;
-	short consolidated_padding_insertion;
-
-	AtomicInfo* last_trak_child_atom;
-	AtomicInfo* moov_atom;
-	AtomicInfo* moov_udta_atom;
-	AtomicInfo* iTunes_list_handler_atom;
-	AtomicInfo* moov_meta_atom;
-	AtomicInfo* file_meta_atom;
-	AtomicInfo* first_mdat_atom;
-	AtomicInfo* first_movielevel_metadata_tagging_atom;
-	AtomicInfo* initial_update_atom;
-	AtomicInfo* first_otiose_freespace_atom;
-	AtomicInfo* padding_store;
-	AtomicInfo* padding_resevoir;
-	FreeAtomListing* first_padding_atom;
-	FreeAtomListing* last_padding_atom;
-};
-
-struct padding_preferences {
-	uint32_t default_padding_size;
-	uint32_t minimum_required_padding_size;
-	uint32_t maximum_present_padding_size;
-};
-
-// Structure that defines the known atoms used by mpeg-4 family of specifications.
-typedef struct {
-  const char*         known_atom_name;
-  const char*					known_parent_atoms[5]; //max known to be tested
-  uint32_t			container_state;
-  int						presence_requirements;
-  uint32_t			box_type;
-} atomDefinition;
-
-typedef struct {
-	uint8_t uuid_form;
-  char* binary_uuid;
-	char* uuid_AP_atom_name;
-} uuid_vitals;
-
-enum {
-  PARENT_ATOM         = 0, //container atom
-	SIMPLE_PARENT_ATOM  = 1,
-	DUAL_STATE_ATOM     = 2, //acts as both parent (contains other atoms) & child (carries data)
-	CHILD_ATOM          = 3, //atom that does NOT contain any children
-	UNKNOWN_ATOM_TYPE   = 4
-};
-
-enum {
-	REQUIRED_ONCE  = 30, //means total of 1 atom per file  (or total of 1 if parent atom is required to be present)
-	REQUIRED_ONE = 31, //means 1 atom per container atom; totalling many per file  (or required present if optional parent atom is present)
-	REQUIRED_VARIABLE = 32, //means 1 or more atoms per container atom are required to be present
-	PARENT_SPECIFIC = 33, //means (iTunes-style metadata) the atom defines how many are present; most are MAX 1 'data' atoms; 'covr' is ?unlimited?
-	OPTIONAL_ONCE = 34, //means total of 1 atom per file, but not required
-	OPTIONAL_ONE = 35, //means 1 atom per container atom but not required; many may be present in a file
-	OPTIONAL_MANY = 36, //means more than 1 occurrence per container atom
-	REQ_FAMILIAL_ONE = OPTIONAL_ONE, //means that one of the family of atoms defined by the spec is required by the parent atom
-	UKNOWN_REQUIREMENTS= 38
-};
-
-enum {
-	SIMPLE_ATOM = 50,
-	VERSIONED_ATOM = 51,
-	EXTENDED_ATOM = 52,
-	PACKED_LANG_ATOM = 53,
-	UNKNOWN_ATOM = 59
-};
-
-enum {
-	PRINT_DATA = 1,
-	PRINT_FREE_SPACE = 2,
-	PRINT_PADDING_SPACE = 4,
-	PRINT_USER_DATA_SPACE = 8,
-	PRINT_MEDIA_SPACE = 16,
-	EXTRACT_ARTWORK = 20,
-	EXTRACT_ALL_UUID_BINARYS = 21
-};
-
-typedef struct {
-  const char*         stik_string;
-  uint8_t       stik_number;
-} stiks;
-
-typedef struct {
-  const char*         storefront_string;
-  uint32_t      storefront_number;
-} sfIDs;
-
-typedef struct {
-	const char* iso639_2_code;
-	const char* iso639_1_code;
-	const char* language_in_english;
-} iso639_lang;
-
-typedef struct {
-  const char* media_rating;
-  const char* media_rating_cli_str;
-} m_ratings;
-
-enum {
-	UNIVERSAL_UTF8,
-	WIN32_UTF16
-};
-
-enum {
- FORCE_M4B_TYPE = 85,
- NO_TYPE_FORCING = 90
-};
-
-enum {
-	FILE_LEVEL_ATOM,
-	MOVIE_LEVEL_ATOM,
-	ALL_TRACKS_ATOM,
-	SINGLE_TRACK_ATOM
-};
-
-enum {
-	UUID_DEPRECATED_FORM,
-	UUID_SHA1_NAMESPACE,
-	UUID_AP_SHA1_NAMESPACE,
-	UUID_OTHER
-};
-
-//==========================================================//
-
+extern atomDefinition KnownAtoms[];
 extern bool parsedfile;
 extern bool file_opened;
 extern bool modified_atoms;
@@ -354,25 +154,7 @@ extern ID3v2FieldDefinition FrameTypeConstructionList[];
 extern ImageFileFormatDefinition ImageList[];
 extern ID3ImageType ImageTypeList[];
 
-//==========================================================//
-
-#ifdef AP_VER
-# define AtomicParsley_version AP_VER
-#else
-# define AtomicParsley_version	"0.9.3"
-#endif
-#define MAX_ATOMS 1024
-#define MAXDATA_PAYLOAD 1256
-
-#define DEFAULT_PADDING_LENGTH          2048;
-#define MINIMUM_REQUIRED_PADDING_LENGTH 0;
-#define MAXIMUM_REQUIRED_PADDING_LENGTH 5000;
-
-//--------------------------------------------------------------------------------------------------------------------------------//
-//--------------------------------------------------------------------------------------------------------------------------------//
-
 void ShowVersionInfo();
-
 void APar_FreeMemory();
 
 short APar_FindParentAtom(int order_in_tree, uint8_t this_atom_level);
@@ -428,21 +210,6 @@ void APar_zlib_inflate(char* in_buffer, uint32_t in_buf_len, char* out_buffer, u
 uint32_t APar_zlib_deflate(char* in_buffer, uint32_t in_buf_len, char* out_buffer, uint32_t out_buf_len);
 
 
-
-/* Declarations of functions and data types used for SHA1 sum
-   library functions.
-   Copyright (C) 2000, 2001, 2003, 2005 Free Software Foundation, Inc.
-*/
-
-typedef struct {
-    uint32_t       time_low;
-    uint16_t       time_mid;
-    uint16_t       time_hi_and_version;
-    uint8_t        clock_seq_hi_and_reserved;
-    uint8_t        clock_seq_low;
-    unsigned char  node[6];
-} ap_uuid_t;
-
 void APar_print_uuid(ap_uuid_t* uuid, bool new_line = true);
 void APar_sprintf_uuid(ap_uuid_t* uuid, char* destination);
 uint8_t APar_uuid_scanf(char* in_formed_uuid, const char* raw_uuid);
@@ -453,23 +220,6 @@ uint8_t APar_extract_uuid_version(ap_uuid_t* uuid, char* binary_uuid_str);
 void APar_generate_uuid_from_atomname(char* atom_name, char* uuid_binary_str);
 void APar_generate_random_uuid(char* uuid_binary_str);
 
-typedef uint32_t md5_uint32;
-
-/* Structure to save state of computation between the single steps.  */
-struct sha1_ctx
-{
-  md5_uint32 A;
-  md5_uint32 B;
-  md5_uint32 C;
-  md5_uint32 D;
-  md5_uint32 E;
-
-  md5_uint32 total[2];
-  md5_uint32 buflen;
-	char buffer[128];   //char buffer[128] __attribute__ ((__aligned__ (__alignof__ (md5_uint32))));
-};
-
-
 /* Initialize structure containing state of computation. */
 extern void sha1_init_ctx (struct sha1_ctx *ctx);
 
@@ -478,14 +228,14 @@ extern void sha1_init_ctx (struct sha1_ctx *ctx);
    starting at BUFFER.
    It is necessary that LEN is a multiple of 64!!! */
 extern void sha1_process_block (const void *buffer, size_t len,
-				struct sha1_ctx *ctx);
+        struct sha1_ctx *ctx);
 
 /* Starting with the result of former calls of this function (or the
    initialization function update the context for the next LEN bytes
    starting at BUFFER.
    It is NOT required that LEN is a multiple of 64.  */
 extern void sha1_process_bytes (const void *buffer, size_t len,
-				struct sha1_ctx *ctx);
+        struct sha1_ctx *ctx);
 
 /* Process the remaining bytes in the buffer and put result from CTX
    in first 20 bytes following RESBUF.  The result is always in little
@@ -526,8 +276,6 @@ int UTF8ToUTF16BE(unsigned char* outb, int outlen, const unsigned char* in, int 
 int UTF16LEToUTF8(unsigned char* out, int outlen, const unsigned char* inb, int inlenb);
 int UTF8ToUTF16LE(unsigned char* outb, int outlen, const unsigned char* in, int inlen);
 
-/*------------- these functions are independent of any secondary licensing -------------*/
-
 int isUTF8(const char* in_string);
 unsigned int utf8_length(const char *in_string, unsigned int char_limit);
 
@@ -535,5 +283,57 @@ int strip_bogusUTF16toRawUTF8 (unsigned char* out, int inlen, wchar_t* in, int o
 
 int test_conforming_alpha_string(char* in_string);
 bool test_limited_ascii(char* in_string, unsigned int str_len);
+
+void APar_ExtractDetails(FILE* isofile, uint8_t optional_output);
+void APar_ExtractBrands(char* filepath);
+
+void printBOM();
+void APar_fprintf_UTF8_data(const char* utf8_encoded_data);
+#if defined (_MSC_VER)
+void APar_unicode_win32Printout(wchar_t* unicode_out, char* utf8_out);
 #endif
 
+void APar_Extract_uuid_binary_file(AtomicInfo* uuid_atom, const char* originating_file, char* output_path);
+void APar_Print_APuuid_atoms(const char *path, char* output_path, uint8_t target_information);
+
+void APar_Print_iTunesData(const char *path, char* output_path, uint8_t supplemental_info, uint8_t target_information, AtomicInfo* ilstAtom = NULL);
+void APar_PrintUserDataAssests(bool quantum_listing = false);
+
+void APar_Extract_ID3v2_file(AtomicInfo* id32_atom, const char* frame_str, const char* originfile, const char* destination_folder, AdjunctArgs* id3args);
+void APar_Print_ID3v2_tags(AtomicInfo* id32_atom);
+
+void APar_Print_ISO_UserData_per_track();
+void APar_Mark_UserData_area(uint8_t track_num, short userdata_atom, bool quantum_listing);
+
+//trees
+void APar_PrintAtomicTree();
+void APar_SimpleAtomPrintout();
+
+uint32_t APar_4CC_CreatorCode(const char* filepath, uint32_t new_type_code);
+void APar_SupplySelectiveTypeCreatorCodes(const char *inputPath, const char *outputPath, uint8_t forced_type_code);
+
+#if defined (DARWIN_PLATFORM)
+bool ResizeGivenImage(const char* filePath, PicPrefs myPicPrefs, char* resized_path);
+#endif
+
+char* GenreIntToString(int genre);
+uint8_t StringGenreToInt(const char* genre_string);
+void ListGenresValues();
+
+stiks* MatchStikString(const char* stik_string);
+stiks* MatchStikNumber(uint8_t in_stik_num);
+void ListStikValues();
+
+sfIDs* MatchStoreFrontNumber(uint32_t storefrontnum);
+
+bool MatchLanguageCode(const char* in_code);
+void ListLanguageCodes();
+
+void ListMediaRatings();
+const char* Expand_cli_mediastring(const char* cli_rating);
+
+char* ID3GenreIntToString(int genre);
+uint8_t ID3StringGenreToInt(const char* genre_string);
+#endif
+
+// vim:ts=2:sw=2:et:
