+++ src/metalist.cpp
@@ -20,11 +20,6 @@
 //==================================================================//
 
 #include "AtomicParsley.h"
-#include "AP_commons.h"
-#include "AP_ID3v2_tags.h"
-#include "AP_arrays.h"
-
-#include "AP_MetadataListings.h"
 
 bool BOM_printed = false;
