+++ src/util.cpp
@@ -26,7 +26,6 @@
 //==================================================================//
 
 #include "AtomicParsley.h"
-#include "AP_commons.h"
 
 
 ///////////////////////////////////////////////////////////////////////////////////////
