+++ src/uuid.cpp
@@ -48,7 +48,6 @@
 //==================================================================//
 
 #include "AtomicParsley.h"
-#include "AP_commons.h"
 
 /*----------------------
 print_hash
